fix: fixed the inserts of some functions // added checks in endponts …

+++ packages/matrix-client-server/src/rooms/room_information/room_tags.ts
@@ -1,3 +1,23 @@
+/*
+ * This file defines handlers for managing room tags in the Matrix client-server API :
+ * https://spec.matrix.org/v1.11/client-server-api/#client-behaviour-15
+ * It includes three main functions:
+ *
+ * 1. `getUserRoomTags`: Retrieves tags associated with a user's room.
+ *
+ * 2. `addUserRoomTag`: Adds a new tag to a user's room.
+ *
+ * 3. `removeUserRoomTag`: Removes a tag from a user's room.
+ *
+ * The only part that is not specified in the Matrix Protocol is the access control logic.
+ * Following Synapse's implementation, we will allow a user to view, add, and remove their tags only.
+ *
+ * For now, it remains possible to add tags to a room you are not part of.
+ *
+ * Maximum lengths:
+ * - `room_tag`: 255 characters
+ */
+
 import type MatrixClientServer from '../../'
 import {
   errMsg,
@@ -15,7 +35,31 @@ export const getUserRoomTags = (
     const userId = (req as Request).params.userId
     const roomId = (req as Request).params.roomId
 
-    clientServer.authenticate(req, res, (data, id) => {
+    // Check if userId and roomId are valid
+    const userIdRegex = /^@[a-zA-Z0-9._=-]+:[a-zA-Z0-9.-]+$/
+    const roomIdRegex = /^![a-zA-Z0-9]+:[a-zA-Z0-9.-]+$/
+
+    if (!userIdRegex.test(userId)) {
+      send(res, 400, errMsg('invalidParam', 'Invalid userId'))
+      return
+    }
+    if (!roomIdRegex.test(roomId)) {
+      send(res, 400, errMsg('invalidParam', 'Invalid roomId'))
+      return
+    }
+
+    clientServer.authenticate(req, res, (token) => {
+      const requesterUserId = token.sub
+
+      if (requesterUserId !== userId) {
+        send(
+          res,
+          403,
+          errMsg('forbidden', 'You are not allowed to view these tags')
+        )
+        return
+      }
+
       clientServer.matrixDb
         .get('room_tags', ['tag', 'content'], {
           user_id: userId,
@@ -69,10 +113,45 @@ export const addUserRoomTag = (
     const roomId = (req as Request).params.roomId
     const _tag = (req as Request).params.tag
 
-    clientServer.authenticate(req, res, (data, id) => {
+    // Check if userId and roomId are valid
+    const userIdRegex = /^@[a-zA-Z0-9._=-]+:[a-zA-Z0-9.-]+$/
+    const roomIdRegex = /^![a-zA-Z0-9]+:[a-zA-Z0-9.-]+$/
+
+    if (!userIdRegex.test(userId)) {
+      send(res, 400, errMsg('invalidParam', 'Invalid userId'))
+      return
+    }
+    if (!roomIdRegex.test(roomId)) {
+      send(res, 400, errMsg('invalidParam', 'Invalid roomId'))
+      return
+    }
+    if (_tag.length > 255) {
+      send(
+        res,
+        400,
+        errMsg('invalidParam', 'The tag must be less than 255 characters')
+      )
+      return
+    }
+
+    clientServer.authenticate(req, res, (token) => {
+      const requesterUserId = token.sub
+      if (requesterUserId !== userId) {
+        send(res, 403, errMsg('forbidden', 'You are not allowed to add tags'))
+        return
+      }
+
       jsonContent(req, res, clientServer.logger, (obj) => {
         validateParameters(res, schema, obj, clientServer.logger, (obj) => {
           const order = obj as { order: number }
+          if (typeof order.order !== 'number' || order.order <= 0) {
+            send(
+              res,
+              400,
+              errMsg('invalidParam', 'The order must be greater than 0')
+            )
+            return
+          }
           clientServer.matrixDb
             .insert('room_tags', {
               user_id: userId,
@@ -103,7 +182,30 @@ export const removeUserRoomTag = (
     const roomId = (req as Request).params.roomId
     const _tag = (req as Request).params.tag
 
-    clientServer.authenticate(req, res, (data, id) => {
+    // Check if userId and roomId are valid
+    const userIdRegex = /^@[a-zA-Z0-9._=-]+:[a-zA-Z0-9.-]+$/
+    const roomIdRegex = /^![a-zA-Z0-9]+:[a-zA-Z0-9.-]+$/
+
+    if (!userIdRegex.test(userId)) {
+      send(res, 400, errMsg('invalidParam', 'Invalid userId'))
+      return
+    }
+    if (!roomIdRegex.test(roomId)) {
+      send(res, 400, errMsg('invalidParam', 'Invalid roomId'))
+      return
+    }
+
+    clientServer.authenticate(req, res, (token) => {
+      const requesterUserId = token.sub
+      if (requesterUserId !== userId) {
+        send(
+          res,
+          403,
+          errMsg('forbidden', 'You are not allowed to remove tags')
+        )
+        return
+      }
+
       clientServer.matrixDb
         .deleteWhere('room_tags', [
           { field: 'user_id', operator: '=', value: userId },
